#56: move transient data and non-general functions out of helpers (#75)
* improve feedback text in /petition

* fix start:dev script

* Moved transient data and non-general functions out of helpers

* fix assign to const error and cleanup
