+++ package.json
@@ -1,7 +1,7 @@
 {
 	"scripts": {
 		"start": "node ./source/bot.js prod",
-		"start:dev": "node source/scripts/uploadCommandsToTestGuild.js && node index.js",
+		"start:dev": "node source/scripts/uploadCommandsToTestGuild.js && node source/bot.js",
 		"clearCommands": "node source/scripts/clearCommands.js",
 		"initialize": "npm ci && node source/scripts/initialize"
 	},
