+++ source/buttons/join.js
@@ -1,12 +1,9 @@
 const Button = require('../classes/Button.js');
 const { guildId } = require('../constants.js');
-const { joinChannel } = require('../helpers.js');
+const { joinChannel } = require('../engines/clubEngine.js');
 
 module.exports = new Button("join",
-	/** Join the club specified in args
-	 * @param {import('discord.js').Interaction} interaction
-	 * @param {string[]} args
-	 */
+	/** Join the club specified in args */
 	(interaction, [channelId]) => {
 		interaction.client.guilds.fetch(guildId).then(guild => {
 			guild.channels.fetch(channelId).then(channel => {
