+++ source/buttons/startevent.js
@@ -1,14 +1,11 @@
 const { GuildScheduledEventStatus, ButtonBuilder, ButtonStyle, ActionRowBuilder } = require('discord.js');
 const Button = require('../classes/Button.js');
-const { getClubDictionary } = require('../helpers.js');
+const { getClubDictionary } = require('../engines/referenceEngine.js');
 const { isModerator } = require("../engines/permissionEngine.js");
 
 const id = "startevent";
 module.exports = new Button(id,
-	/** Start a club's event
-	 * @param {import('discord.js').Interaction} interaction
-	 * @param {Array<string>} args
-	 */
+	/** Start a club's event */
 	(interaction, [eventId]) => {
 		const { hostId } = getClubDictionary()[interaction.message.channel.id];
 		if (!isModerator(interaction.member) && interaction.user.id !== hostId) {
