+++ source/commands/club-invite.js
@@ -1,9 +1,8 @@
-const { ButtonBuilder, ActionRowBuilder } = require('@discordjs/builders');
-const { ButtonStyle } = require('discord.js');
+const { ButtonBuilder, ActionRowBuilder, ButtonStyle } = require('discord.js');
 const Command = require('../classes/Command.js');
-const { clubEmbedBuilder } = require('../engines/messageEngine.js');
-const { getClubDictionary } = require('../helpers.js');
 const { SAFE_DELIMITER } = require('../constants.js');
+const { clubEmbedBuilder } = require('../engines/messageEngine.js');
+const { getClubDictionary } = require('../engines/referenceEngine.js');
 
 const options = [
 	// can't use channel mention because users can't mention channels that are invisible to them (even by constructing the mention manually)
