+++ source/commands/petition.js
@@ -1,5 +1,5 @@
 const Command = require('../classes/Command.js');
-const { checkPetition, getTopicNames } = require('../engines/channelEngine.js');
+const { checkPetition, getTopicNames } = require('../engines/referenceEngine.js');
 
 const options = [
 	{ type: "String", name: "topic-name", description: "Make sure the topic doesn't already exist", required: true, choices: [] }
@@ -13,7 +13,7 @@ module.exports.execute = (interaction) => {
 	const topicName = interaction.options.getString("topic-name").toLowerCase();
 	if (!getTopicNames().includes(topicName)) {
 		checkPetition(interaction.guild, topicName, interaction.user);
-		interaction.reply(`Your petition for **${topicName}** has been recorded!`)
+		interaction.reply(`Your petition for a **${topicName}** text channel has been recorded!`)
 			.catch(console.error);
 	} else {
 		interaction.reply({ content: `A channel for ${topicName} already exists, you can use the Channels & Roles Browser to make it visible.`, ephemeral: true })
