+++ source/commands/pin-list.js
@@ -1,5 +1,6 @@
 const Command = require('../classes/Command.js');
-const { pinList } = require('../helpers.js');
+const { saveObject } = require('../helpers.js');
+const { listMessages, buildListMessagePayload } = require('../engines/referenceEngine.js');
 
 const id = "pin-list";
 const options = [
@@ -13,7 +14,17 @@ module.exports = new Command(id, "Pin the petition or club list message in this 
  */
 module.exports.execute = (interaction) => {
 	const listType = interaction.options.getString("list-type").toLowerCase();
-	pinList(interaction.guild.channels, interaction.channel, listType);
+	buildListMessagePayload(interaction.guild.memberCount, listType).then(messageOptions => {
+		interaction.channel.send(messageOptions).then(message => {
+			listMessages[listType] = {
+				"messageId": message.id,
+				"channelId": message.channelId
+			}
+			saveObject(listMessages, "listMessageIds.json");
+			message.pin();
+		})
+	}).catch(console.error);
+
 	interaction.reply({ content: `Pinning the ${listType} list succeded.`, ephemeral: true })
 		.catch(console.error);
 }
