+++ source/modalSubmissions/changeclubseats.js
@@ -1,14 +1,11 @@
-const { Interaction } = require('discord.js');
 const ModalSubmission = require('../classes/ModalSubmission.js');
 const { clubEmbedBuilder } = require('../engines/messageEngine.js');
-const { getClubDictionary, updateClub, updateClubDetails, updateList } = require('../helpers.js');
+const { updateClubDetails } = require('../engines/clubEngine.js');
+const { getClubDictionary, updateClub, updateList } = require('../engines/referenceEngine.js');
 
 const id = "changeclubseats";
 module.exports = new ModalSubmission(id,
-	/** Set the max members and isRecruting for the club with provided id
-	 * @param {Interaction} interaction
-	 * @param {Array<string>} args
-	 */
+	/** Set the max members and isRecruting for the club with provided id */
 	async (interaction, [clubId]) => {
 		const club = getClubDictionary()[clubId];
 		const { fields } = interaction;
